requirements.txt
asgiref==3.2.7
astroid==2.0.4
asttokens==2.0.5
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
compound-word-splitter==0.4
cwsplit==0.4.1
Django==3.0.4
executing==0.6.0
icecream==2.1.1
idna==2.10
importlib-metadata==4.6.1
isort==4.3.4
jsonpickle==2.0.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
mercurial==5.2
numpy==1.21.0
pandas==1.2.5
pyenchant==3.2.0
Pygments==2.9.0
pylint==2.1.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.4.1
requests==2.25.1
six==1.11.0
soupsieve==2.2.1
sqlparse==0.3.1
typing-extensions==3.10.0.0
urllib3==1.26.6
wrapt==1.10.11
zipp==3.5.0
